Rename 'run_tests' to 'test' and update references
The codebase has been refactored to rename the 'run_tests' function and files to simply 'test'. This change includes all references and imports throughout the entire code, ensuring code consistency and improved readability.
